requirements.txt
flask
werkzeug
python-docx
PyPDF2